+++ tests/test_git_repo_scanner.py
@@ -184,8 +184,12 @@ class ChunkGeneratorTests(ScannerTestCase):
     ):
         self.git_options.fetch = False
         self.git_options.branch = "bar"
+        mock_foo = mock.MagicMock()
+        mock_foo.name = "foo"
+        mock_bar = mock.MagicMock()
+        mock_bar.name = "bar"
         mock_repo.return_value.remotes.origin.fetch.return_value = ["foo", "bar"]
-        mock_repo.return_value.branches = ["foo", "bar"]
+        mock_repo.return_value.branches = [mock_foo, mock_bar]
         test_scanner = scanner.GitRepoScanner(
             self.global_options, self.git_options, "."
         )
@@ -193,7 +197,9 @@ class ChunkGeneratorTests(ScannerTestCase):
         for _ in test_scanner.chunks:
             pass
         mock_repo.return_value.remotes.origin.fetch.assert_not_called()
-        mock_iter_commits.assert_has_calls((mock.call(mock_repo.return_value, "bar"),))
+        mock_iter_commits.assert_has_calls(
+            (mock.call(mock_repo.return_value, mock_bar),)
+        )
 
     @mock.patch("tartufo.scanner.GitRepoScanner._iter_branch_commits")
     @mock.patch("git.Repo")
@@ -202,8 +208,12 @@ class ChunkGeneratorTests(ScannerTestCase):
     ):
         self.git_options.fetch = True
         self.git_options.branch = "bar"
+        mock_foo = mock.MagicMock()
+        mock_foo.name = "foo"
+        mock_bar = mock.MagicMock()
+        mock_bar.name = "bar"
         mock_repo.return_value.remotes.origin.fetch.return_value = ["foo", "bar"]
-        mock_repo.return_value.branches = ["foo", "bar"]
+        mock_repo.return_value.branches = [mock_foo, mock_bar]
         test_scanner = scanner.GitRepoScanner(
             self.global_options, self.git_options, "."
         )
@@ -211,7 +221,9 @@ class ChunkGeneratorTests(ScannerTestCase):
         for _ in test_scanner.chunks:
             pass
         mock_repo.return_value.remotes.origin.fetch.assert_called()
-        mock_iter_commits.assert_has_calls((mock.call(mock_repo.return_value, "bar"),))
+        mock_iter_commits.assert_has_calls(
+            (mock.call(mock_repo.return_value, mock_bar),)
+        )
 
     @mock.patch("tartufo.scanner.GitRepoScanner._iter_branch_commits")
     @mock.patch("tartufo.scanner.GitRepoScanner._iter_diff_index")
@@ -381,7 +393,8 @@ class IterBranchCommitsTests(ScannerTestCase):
         commits = list(test_scanner._iter_branch_commits(mock_repo, mock_branch))
         # Because "since commit" is exclusive, only the 2 commits before it are ever yielded
         self.assertEqual(
-            commits, [(mock_commit_2, mock_commit_1)],
+            commits,
+            [(mock_commit_2, mock_commit_1)],
         )