+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Features:
 
 * #42 - Report output on clean or successful scan. Add new `-q/--quiet` option to suppress output
 * #43 - Report out of the list of exclusions. Add new `-v/--verbose` option to print exclusions
+* Updated BFG refs from 1.13.0 to 1.13.2
 
 
 v2.2.1 - 02 December 2020